Add DocumentHelper and move cleaned up RemoveInternalDocumentValues there
